Se agrega el alert y la funcionalidad de verificar la palabra

+++ src/main/java/com/example/juegodeahorcado/controller/WelcomeController.java
@@ -1,5 +1,6 @@
 package com.example.juegodeahorcado.controller;
 
+import com.example.juegodeahorcado.view.alert.AlertBox;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.scene.control.TextField;
@@ -7,7 +8,9 @@ import com.example.juegodeahorcado.model.Player;
 import com.example.juegodeahorcado.view.GameStage;
 import com.example.juegodeahorcado.view.WelcomeStage;
 
+
 import java.io.IOException;
+import java.util.regex.Pattern;
 
 public class WelcomeController {
     @FXML
@@ -17,8 +20,18 @@ public class WelcomeController {
     void onHandleButtonPlay(ActionEvent event) throws IOException {
         String palabraClave = campoDeTextoPalabraClave.getText();
 
-        Player player = new Player(1, palabraClave);
-        GameStage.getInstance().getGameController().setPlayer(player);
-        WelcomeStage.deleteInstance();
+        if (validarPalabraClave(palabraClave)) {
+            Player player = new Player(1, palabraClave);
+            GameStage.getInstance().getGameController().setPlayer(player);
+            WelcomeStage.deleteInstance();
+        } else {
+            String mensaje = "Por favor, ingresa únicamente palabras en el campo de texto.";
+            AlertBox.showMessage("Error", "Palabra Clave Inválida", mensaje);
+        }
+    }
+
+    private boolean validarPalabraClave(String palabraClave) {
+        // Utilizamos una expresión regular para verificar si solo se han utilizado palabras
+        return Pattern.matches("\\b[a-zA-Z]+\\b", palabraClave);
     }
 }
